Replace RainElement enum with newtype struct
After our requirements were refined, we now know that the rain must be
represented as floating point number, so there's no point in keeping it
as an Enum.

Rather, we use a newtype here, to restrict the interface of the
rain-representing type to a bare minimum in what can be done with it.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>