Remove impl for element simulation
After refinement of the requirements, it is now clear that the levels
are represented as floating point numbers.

Thus, a landscape like:

    [ 2, 1, 1 ]

with one hour of rain would end in

    [ 2.33, 2.33, 2.33 ]

so the implementation must take care of shipping water around from the
first element to the third in this landscape, for example.

Thus, the implementation must be rewritten to be more "recursive"
(although because we are implementing a simulation here, it can be
iterative, actually).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>